QFG1EGA: Don't preset stats at startup
One of my additions to QFG1EGA was the option to start at any room you wanted. This also set all stats to 80.

As it turns out, creating a character doesn't zero out all stats. Three spells are still remembered. So we just keep all stats at zero.

If you'd like to set all stats to 80, just press Alt-X, as is the case in the original debug script.